back-end/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
# sqlite3
python-dotenv
passlib
PyJWT
email-validator
migrations
alembic
bcrypt